+++ tests/integration/components/yeti-table/general-test.js
@@ -517,7 +517,7 @@ module('Integration | Component | yeti-table (general)', function(hooks) {
         <table.body/>
 
         <div id="totalColumns">{{table.totalColumns.length}}</div>
-        <div id="visibleColumns">{{table.visibleColumns}}</div>
+        <div id="visibleColumns">{{table.visibleColumns.length}}</div>
         <div id="totalRows">{{table.totalRows}}</div>
         <div id="visibleRows">{{table.visibleRows.length}}</div>