+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ## v1.0.0
 
 - [BREAKING CHANGE] `table.visibleRows` is no longer a number of visible rows but the rows data itself. If you want the number of visible rows, you can use `{{table.visibleRows.length}}`
-- [BREAKING CHANGE] `table.totalColumns` is no longer a number of total columns but an array of object that represent the columns. If you want the number of total columns, you can use `{{table.totalColumns.length}}`
+- [BREAKING CHANGE] `table.totalColumns` is no longer a number of total columns but an array of objects that represent the columns. If you want the number of total columns, you can use `{{table.totalColumns.length}}`
+- [BREAKING CHANGE] `table.visibleColumns` is no longer a number of visible columns but an array of objects that represent the visible columns. If you want the number of visible columns, you can use `{{table.visibleColumns.length}}`
 - [CHORE] updated dependencies and removed `@argument` decorator because it was stopping us from supporting latest ember versions